refactor: changed filenames to *.js
