+++ UTILS/SF6-data/characterData.js
