+++ src/scraper/characterData/characterSkills.js
