+++ src/scraper/commandListPage/commandScrape.js
