+++ src/scraper/frameDataPage/vitalityValue.js
