+++ src/scraper/helperFunctions/_handleImage.js
