+++ src/scraper/helperFunctions/formatForDataBase.js
